+++ 06_gpu_and_ml/yolo/demo_images/bees.png
